requirements.txt
aiohttp==3.8.5
python-dateutil==2.8.2
python-dotenv==1.0.0 